+++ FanFollow/FanFollow/Source/NetworkManager/NetworkManager.swift
@@ -8,6 +8,12 @@ import Foundation
 import RxSwift
 
 final class NetworkManager: Network {
+    private let session: URLSession
+    
+    init(session: URLSession = URLSession.shared) {
+        self.session = session
+    }
+    
     func response(_ request: URLRequest) -> Observable<(response: URLResponse, data: Data)> {
         return Observable.create { emitter in
             let task = URLSession.shared.dataTask(with: request) { data, response, error in
