+++ FanFollow/FanFollow/Source/NetworkManager/Protocol/URLSessionDataTaskType.swift
@@ -1,0 +1,11 @@
+//
+//  URLSessionDataTaskType.swift
+//  FanFollow
+//
+//  Copyright (c) 2023 Minii All rights reserved.
+
+import Foundation
+
+protocol URLSessionDataTaskType {
+    func resume()
+}
