+++ FanFollow/FanFollow/Source/NetworkManager/Protocol/URLSessionType.swift
@@ -1,0 +1,14 @@
+//
+//  URLSessionType.swift
+//  FanFollow
+//
+//  Copyright (c) 2023 Minii All rights reserved.
+
+import Foundation
+
+protocol URLSessionType {
+    func dataTask(
+        with request: URLRequest,
+        completionHandler: @escaping @Sendable (Data?, URLResponse?, Error?) -> Void
+    ) -> URLSessionDataTaskType
+}
